+++ modules/farm_rothamsted_experiment_research/src/Entity/RothamstedDesign.php
@@ -31,10 +31,12 @@ use Drupal\user\UserInterface;
  *       "add" = "Drupal\farm_rothamsted_experiment_research\Form\DesignEntityForm",
  *       "edit" = "Drupal\farm_rothamsted_experiment_research\Form\DesignEntityForm",
  *       "delete" = "Drupal\Core\Entity\ContentEntityDeleteForm",
+ *       "entity-status-action-form" = "Drupal\farm_rothamsted_experiment_research\Form\EntityStatusChangeActionForm",
  *     },
  *     "route_provider" = {
  *       "default" = "Drupal\entity\Routing\AdminHtmlRouteProvider",
  *       "revision" = "\Drupal\entity\Routing\RevisionRouteProvider",
+ *       "status-change" = "Drupal\farm_rothamsted_experiment_research\Routing\EntityStatusChangeRouteProvider",
  *     },
  *     "local_task_provider" = {
  *       "default" = "\Drupal\farm_ui_menu\Menu\DefaultSecondaryLocalTaskProvider",
@@ -69,6 +71,7 @@ use Drupal\user\UserInterface;
  *     "version-history" = "/rothamsted/design/{rothamsted_design}/revisions",
  *     "revision" = "/rothamsted/design/{rothamsted_design}/revisions/{rothamsted_design_revision}/view",
  *     "revision-revert-form" = "/rothamsted/design/{rothamsted_design}/revisions/{rothamsted_design_revision}/revert",
+ *     "entity-status-action-form" = "/rothamsted/design/change-status"
  *   },
  * )
  */