+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Enable farmOS-map snapshot behavior. [#669](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/669)
 - Add Study ID field to Proposals. [#662](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/662)
+- Add Study ID filter and column to proposal view. [#773](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/773)
 
 ### Fixed